Various changes:
  * Fixed Bug 63 and 64.

  * Handled nonstandard sed location.

  * Used for-loop in man page install.

  * Included casts for 64-bit builds.

  * Added placeholder for FreeBSD options.

  * Included unistd.h for getopt() and stdlib.h for malloc().

  * Implemented SoText[23] workaround for glibc-2.2.* iconv().

  * Split long lines in SoHandleBoxDraggerGeom.h and
    SoTransformerDraggerGeom.h in lib/interaction/src/draggers/geom.

  * Added IV_NO_OVERLAYS/OIV_NO_OVERLAYS variables to disable overlay planes.

+++ apps/demos/qmorf/qmorf.c++
@@ -43,12 +43,10 @@
 // Written by Gavin Bell for Silicon Graphics
 //
 #include <stdio.h>
-#include <getopt.h>
 #include <math.h>
 #include <stdlib.h>
 #include <unistd.h>
 #include <string.h>
-#include <unistd.h>
 
 
 #include <Inventor/Xt/SoXt.h>
@@ -576,7 +574,7 @@ buildUI( Widget appWindow,
     w[2] = XmCreateToggleButtonGadget(form, "colorAnimate", resources,
 				      n);
     XtAddCallback(w[2], XmNvalueChangedCallback,
-		  toggleAnimation, (XtPointer)COLOR);
+		  toggleAnimation, (XtPointer)(unsigned long)COLOR);
     n = 0;
     XtSetArg(resources[n], XmNlabelString, STRING("Animate")); ++n;
     XtSetArg(resources[n], XmNrightAttachment, XmATTACH_FORM); ++n;
@@ -586,7 +584,7 @@ buildUI( Widget appWindow,
     w[5] = XmCreateToggleButtonGadget(form, "shapeAnimate", resources,
 				      n);
     XtAddCallback(w[5], XmNvalueChangedCallback,
-		  toggleAnimation, (XtPointer)SHAPE);
+		  toggleAnimation, (XtPointer)(unsigned long)SHAPE);
     n = 0;
 
     XtSetArg(resources[n], XmNlabelString, STRING("Quit")); ++n;
